+++ crates/cargo-contract/src/cmd/extrinsics/integration_tests.rs
@@ -38,6 +38,16 @@ fn cargo_contract(path: &Path) -> assert_cmd::Command {
     cmd
 }
 
+// Find the contract address in the output
+fn extract_contract_address(stdout: &str) -> &str {
+    let regex = regex::Regex::new("Contract ([0-9A-Za-z]+)").unwrap();
+    let caps = regex
+        .captures(stdout)
+        .expect("contract account regex capture");
+    let contract_account = caps.get(1).unwrap().as_str();
+    contract_account
+}
+
 /// Spawn and manage an instance of a compatible contracts enabled chain node.
 #[allow(dead_code)]
 struct ContractsNodeProcess {
@@ -169,6 +179,7 @@ async fn build_upload_instantiate_call() {
     let output = cargo_contract(project_path.as_path())
         .arg("upload")
         .args(["--suri", "//Alice"])
+        .arg("-x")
         .output()
         .expect("failed to execute process");
     let stderr = str::from_utf8(&output.stderr).unwrap();
@@ -179,18 +190,14 @@ async fn build_upload_instantiate_call() {
         .args(["--constructor", "new"])
         .args(["--args", "true"])
         .args(["--suri", "//Alice"])
+        .arg("-x")
         .output()
         .expect("failed to execute process");
     let stdout = str::from_utf8(&output.stdout).unwrap();
     let stderr = str::from_utf8(&output.stderr).unwrap();
     assert!(output.status.success(), "instantiate failed: {stderr}");
 
-    // find the contract address in the output
-    let regex = regex::Regex::new("Contract ([0-9A-Za-z]+)").unwrap();
-    let caps = regex
-        .captures(stdout)
-        .expect("contract account regex capture");
-    let contract_account = caps.get(1).unwrap().as_str();
+    let contract_account = extract_contract_address(stdout);
     assert_eq!(48, contract_account.len(), "{stdout:?}");
 
     let call_get_rpc = |expected: bool| {
@@ -199,7 +206,6 @@ async fn build_upload_instantiate_call() {
             .args(["--message", "get"])
             .args(["--contract", contract_account])
             .args(["--suri", "//Alice"])
-            .arg("--dry-run")
             .assert()
             .stdout(predicate::str::contains(expected.to_string()));
     };
@@ -211,6 +217,7 @@ async fn build_upload_instantiate_call() {
         .args(["--message", "flip"])
         .args(["--contract", contract_account])
         .args(["--suri", "//Alice"])
+        .arg("-x")
         .assert()
         .stdout(predicate::str::contains("ExtrinsicSuccess"));
 
@@ -252,6 +259,7 @@ async fn build_upload_remove() {
     let output = cargo_contract(project_path.as_path())
         .arg("upload")
         .args(["--suri", "//Alice"])
+        .arg("-x")
         .output()
         .expect("failed to execute process");
     let stderr = str::from_utf8(&output.stderr).unwrap();
@@ -316,6 +324,7 @@ async fn build_upload_instantiate_info() {
     let output = cargo_contract(project_path.as_path())
         .arg("upload")
         .args(["--suri", "//Alice"])
+        .arg("-x")
         .output()
         .expect("failed to execute process");
     let stderr = str::from_utf8(&output.stderr).unwrap();
@@ -326,18 +335,14 @@ async fn build_upload_instantiate_info() {
         .args(["--constructor", "new"])
         .args(["--args", "true"])
         .args(["--suri", "//Alice"])
+        .arg("-x")
         .output()
         .expect("failed to execute process");
     let stdout = str::from_utf8(&output.stdout).unwrap();
     let stderr = str::from_utf8(&output.stderr).unwrap();
     assert!(output.status.success(), "instantiate failed: {stderr}");
 
-    // find the contract address in the output
-    let regex = regex::Regex::new("Contract ([0-9A-Za-z]+)").unwrap();
-    let caps = regex
-        .captures(stdout)
-        .expect("contract account regex capture");
-    let contract_account = caps.get(1).unwrap().as_str();
+    let contract_account = extract_contract_address(stdout);
     assert_eq!(48, contract_account.len(), "{stdout:?}");
 
     cargo_contract(project_path.as_path())
@@ -347,6 +352,18 @@ async fn build_upload_instantiate_info() {
         .expect("failed to execute process");
     let stderr = str::from_utf8(&output.stderr).unwrap();
     assert!(output.status.success(), "getting info failed: {stderr}");
+
+    cargo_contract(project_path.as_path())
+        .arg("info")
+        .args(["--contract", contract_account])
+        .arg("--output-json")
+        .output()
+        .expect("failed to execute process");
+    let stderr = str::from_utf8(&output.stderr).unwrap();
+    assert!(
+        output.status.success(),
+        "getting info as JSON format failed: {stderr}"
+    );
 
     // prevent the node_process from being dropped and killed
     let _ = node_process;